Add files via upload
Folders
Vcode is a Microsoft Visual studio debugger launcher setting for the browser.
Libraries all libraries used to turn the app. P5js, ml5, cLmtrackr.js was used on a different function but not much.
Index.html file which contains the link libraries live or local,  jsconfig.json is a Microsoft Visual Studio code configuration file for the browser and port setting if developing on the local machine. Sketch.js is the P5js file that holds the DoubletapAi code and Style.css for style sheet.